refactor: move interaction response methods to interface
